tools/requirements.txt
paddleocr
fpdf
pillow
